requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-google-genai==0.0.5
google-generativeai==0.3.2